node: Extract StreamPublisher and StreamSubscription
Move the publisher and subscription types into separate modules. This is
a step towards thinning-out the stream module which has grown very
large.
