+++ p2panda/src/streams/mod.rs
@@ -5,8 +5,10 @@ mod ephemeral_stream;
 mod event_stream;
 mod external_stream;
 mod local_stream;
+mod publisher;
 mod replay;
 mod stream;
+mod subscription;
 mod sync_metrics;
 
 use p2panda_core::Topic;
@@ -26,12 +28,11 @@ pub use event_stream::SystemEvent;
 pub(crate) use event_stream::event_stream;
 pub use external_stream::ExternalStreamFuture;
 pub(crate) use local_stream::LocalStreamFuture;
+pub use publisher::{ImportError, PublishError, PublishFuture, StreamPublisher};
 pub use replay::{ReplayError, StreamFrom};
 pub(crate) use stream::processed_stream;
-pub use stream::{
-    ImportError, ProcessedOperation, PublishError, PublishFuture, Source, StreamEvent,
-    StreamPublisher, StreamSubscription,
-};
+pub use stream::{ProcessedOperation, Source, StreamEvent};
+pub use subscription::StreamSubscription;
 pub use sync_metrics::{SessionPhase, SyncError};
 
 pub(crate) type Event = crate::processor::Event<LogId, Extensions, Topic>;
