+++ p2panda/src/streams/stream.rs
@@ -2,14 +2,11 @@
 
 use std::borrow::Borrow;
 use std::fmt::Debug;
-use std::marker::PhantomData;
-use std::pin::Pin;
 use std::sync::Arc;
-use std::task::{Context, Poll};
 
+use futures_util::StreamExt;
 use futures_util::stream::BoxStream;
-use futures_util::{FutureExt, Stream, StreamExt};
-use p2panda_core::cbor::{DecodeError, EncodeError, decode_cbor, encode_cbor};
+use p2panda_core::cbor::{DecodeError, decode_cbor};
 use p2panda_core::traits::Digest;
 use p2panda_core::{Hash, Topic, VerifyingKey};
 use p2panda_net::NodeId;
@@ -19,27 +16,27 @@ use p2panda_net::sync::SyncHandle;
 use p2panda_net::utils::ShortFormat;
 use p2panda_spaces::SpaceEvent;
 use p2panda_store::SqliteStore;
-use p2panda_store::operations::OperationStore;
 use p2panda_stream::spaces::SpacesResult;
 use p2panda_sync::protocols::TopicLogSyncEvent;
 use serde::{Deserialize, Serialize};
-use thiserror::Error;
 use tokio::sync::{mpsc, oneshot};
 use tokio_stream::wrappers::ReceiverStream;
 use tracing::{debug, warn};
 
-use crate::forge::{Forge, ForgeError, OperationForge};
+use crate::forge::OperationForge;
 use crate::node::{AckPolicy, CreateStreamError};
-use crate::operation::{Extensions, Header, LogId, Operation};
+use crate::operation::{Extensions, Header, Operation};
 use crate::processor::{ProcessorError, ProcessorStatus};
+use crate::spaces::spawn_repair_task;
 use crate::spaces::types::SpacesManager;
-use crate::spaces::{RepairError, spawn_repair_task};
 use crate::streams::acked::{Acked, AckedError};
 use crate::streams::external_stream::{
     ExternalStream, ExternalStreamEvent, ExternalStreamFuture, SessionId,
 };
 use crate::streams::local_stream::{LocalStream, LocalStreamEvent, LocalStreamFuture};
+use crate::streams::publisher::StreamPublisher;
 use crate::streams::replay::{ReplayError, StreamFrom, replay_log_ranges};
+use crate::streams::subscription::StreamSubscription;
 use crate::streams::sync_metrics::{self, Aggregator, SessionPhase, SyncError};
 use crate::streams::{Event, Pipeline};
 
@@ -404,7 +401,7 @@ where
         });
     }
 
-    let tx = StreamPublisher {
+    let tx = StreamPublisher::new(
         topic,
         forge,
         publish_tx,
@@ -412,15 +409,9 @@ where
         import_local_tx,
         repair_tx,
         to_output_tx,
-        _marker: PhantomData,
-    };
+    );
 
-    let rx = StreamSubscription {
-        topic,
-        store,
-        acked,
-        stream: ReceiverStream::new(app_rx),
-    };
+    let rx = StreamSubscription::new(topic, store, acked, ReceiverStream::new(app_rx));
 
     Ok((tx, rx))
 }
@@ -610,303 +601,6 @@ where
     }
 
     Some(forward_events)
-}
-
-/// Publish messages into a topic stream.
-///
-/// Any message type `M` can be published as long as it can be encoded into bytes by implementing
-/// serde's [`Serialize`] and [`Deserialize`] traits.
-///
-/// ## Example
-///
-/// ```rust
-/// # use p2panda_core::Topic;
-/// # use serde::{Serialize, Deserialize};
-/// # #[tokio::main]
-/// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
-/// # let node = p2panda::builder().spawn().await?;
-/// #
-/// let our_trees = Topic::random();
-///
-/// #[derive(Clone, Debug, Serialize, Deserialize)]
-/// struct Tree {
-///     leafy: bool,
-///     latin_name: String,
-/// }
-///
-/// let (tx, _rx) = node.stream::<Tree>(our_trees).await?;
-///
-/// tx.publish(Tree {
-///     leafy: true,
-///     latin_name: "Acer pseudoplatanus".into(),
-/// }).await?;
-/// #
-/// # Ok(())
-/// # }
-/// ```
-///
-/// ## Append-only log
-///
-/// The Node API internally maintains an append-only log data-type for published application
-/// messages in a topic stream.
-///
-/// Publishing a message creates and signs an [`Operation`] which gets automatically appended to
-/// the author's log for the given topic. The message itself is the payload of the created
-/// operation.
-///
-/// ```plain
-/// Author "Panda"
-/// Topic "Trees" Log: [ Header ] <-- [ Header ] <-- [ Header ] ...
-///                         |             |              |
-///                         v            ...            ...
-///                     [ Body ]
-///               "Acer pseudoplatanus"
-///
-/// Author "Icebear"
-/// Topic "Trees" Log: [ Header ] <-- [ Header ] ...
-///                         |             |
-///                         v            ...
-///                     [ Body ]
-///                 "Pinus halepensis"
-/// ```
-///
-/// ## External sources
-///
-/// Operations can be imported from external sources into the processing pipeline by calling
-/// [`StreamPublisher::import`]. This allows transporting data via sneakernets (USB stick, etc.) or
-/// other sync solutions.
-#[derive(Clone, Debug)]
-pub struct StreamPublisher<M> {
-    topic: Topic,
-    forge: OperationForge,
-    #[allow(clippy::type_complexity)]
-    pub(crate) publish_tx: mpsc::Sender<(Operation, Option<M>, oneshot::Sender<Event>)>,
-    import_external_tx: mpsc::Sender<(
-        BoxStream<'static, Operation>,
-        oneshot::Sender<ExternalStreamFuture>,
-    )>,
-    #[allow(clippy::type_complexity)]
-    import_local_tx: mpsc::Sender<(
-        BoxStream<'static, Operation>,
-        oneshot::Sender<LocalStreamFuture>,
-    )>,
-    pub(crate) to_output_tx: mpsc::Sender<Vec<StreamEvent<M>>>,
-    pub(crate) repair_tx: mpsc::Sender<oneshot::Sender<Result<bool, RepairError>>>,
-    _marker: PhantomData<M>,
-}
-
-impl<M> StreamPublisher<M>
-where
-    M: Serialize,
-{
-    /// Associated topic.
-    pub fn topic(&self) -> Topic {
-        self.topic
-    }
-
-    /// Publish a message into a topic stream.
-    ///
-    /// Locally created operations are processed by the same pipeline as remotely received
-    /// operations. It is possible to await the processing result which can be useful for some
-    /// applications if they want to block UI components etc.
-    pub async fn publish(&self, message: M) -> Result<PublishFuture, PublishError> {
-        self.publish_inner(Some(message), false).await
-    }
-
-    /// Deletes all our previously published messages in this topic stream.
-    ///
-    /// This signals to all other nodes that they should remove them as well.
-    ///
-    /// A message can be optionally added when pruning, allowing to publish a "snapshot" /
-    /// state-based CRDT of the current state, so nodes can still consistently re-create all state,
-    /// even if previous messages are gone.
-    ///
-    /// Internally we're applying append-only log prefix deletion, meaning that the log's prefix
-    /// gets pruned. The prefix is the set of operations in the log's sequence which are causally
-    /// "older" / before the point where the prune flag was set.
-    pub async fn prune(&self, message: Option<M>) -> Result<PublishFuture, PublishError> {
-        self.publish_inner(message, true).await
-    }
-
-    /// Import an external source of operations.
-    ///
-    /// Please note: Operations do not contain any information by themselves about to which topic
-    /// they belong. By importing operations into a topic stream, they will be assigned to this
-    /// topic. Make sure you accordingly routed operations into the correct topic before.
-    pub async fn import(
-        &self,
-        stream: impl Stream<Item = Operation> + Send + 'static,
-    ) -> Result<ExternalStreamFuture, ImportError> {
-        // Send stream to processor.
-        let stream = Box::pin(stream);
-        let (ready_tx, ready_rx) = oneshot::channel::<ExternalStreamFuture>();
-        self.import_external_tx
-            .send((stream, ready_tx))
-            .await
-            .map_err(|err| ImportError::SendToProcessor(err.to_string()))?;
-
-        // Await receiving the session id and future which will complete when the external stream
-        // closes and all operations have been processed.
-        ready_rx
-            .await
-            .map_err(|err| ImportError::ReceiveFromProcessor(err.to_string()))
-    }
-
-    /// Import a local source of operations.
-    ///
-    /// This method can be used for importing some locally forged operations in a batch. The user
-    /// will receive no "import" events on the stream, only events resulting from publishing the
-    /// operations themselves.
-    pub(crate) async fn import_local(
-        &self,
-        stream: impl Stream<Item = Operation> + Send + 'static,
-    ) -> Result<LocalStreamFuture, ImportError> {
-        // Send stream to processor.
-        let stream = Box::pin(stream);
-        let (ready_tx, ready_rx) = oneshot::channel::<LocalStreamFuture>();
-        self.import_local_tx
-            .send((stream, ready_tx))
-            .await
-            .map_err(|err| ImportError::SendToProcessor(err.to_string()))?;
-
-        // Await receiving the future which will complete when the stream
-        // closes and all operations have been processed.
-        ready_rx
-            .await
-            .map_err(|err| ImportError::ReceiveFromProcessor(err.to_string()))
-    }
-
-    async fn publish_inner(
-        &self,
-        message: Option<M>,
-        prune_flag: bool,
-    ) -> Result<PublishFuture, PublishError> {
-        // Create, sign and persist operation with given payload.
-        let extensions = Extensions::builder(LogId::from_topic(self.topic()))
-            .prune_flag(prune_flag)
-            .build();
-
-        let body_bytes = match message {
-            Some(ref message) => Some(encode_cbor(&message)?),
-            None => None,
-        };
-
-        let operation = self
-            .forge
-            .create_operation(
-                Some(self.topic()),
-                extensions.log_id(),
-                body_bytes,
-                extensions,
-            )
-            .await?;
-        let hash = operation.hash;
-
-        // Start processing operation in pipeline. Keep a oneshot receiver around to allow users to
-        // optionally await & get informed when processing has finished.
-        let (processed_tx, processed_rx) = oneshot::channel();
-        self.publish_tx
-            .send((operation.clone(), message, processed_tx))
-            .await
-            .map_err(|err| PublishError::SendToProcessor(err.to_string()))?;
-
-        Ok(PublishFuture { hash, processed_rx })
-    }
-}
-
-/// Future which can be awaited to find out when a locally published operation has finished
-/// processing.
-#[derive(Debug)]
-pub struct PublishFuture {
-    hash: Hash,
-    processed_rx: oneshot::Receiver<Event>,
-}
-
-impl PublishFuture {
-    /// Returns hash of the published operation.
-    pub fn hash(&self) -> Hash {
-        self.hash
-    }
-}
-
-impl Future for PublishFuture {
-    type Output = Result<Event, oneshot::error::RecvError>;
-
-    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        self.processed_rx.poll_unpin(cx)
-    }
-}
-
-/// Subscription to events arriving from a topic stream.
-///
-/// A topic stream emits:
-///
-/// - Locally created or remotely synced [`ProcessedOperation`] with application messages inside
-/// - Topic-scoped system-events, for example if a sync session has begun and how much will be sent
-/// - Critical errors such as [`AckedError`] coming from the processing pipeline
-///
-/// ## Example
-///
-/// ```no_run
-/// use futures_util::StreamExt;
-/// use p2panda_core::Topic;
-/// # #[tokio::main]
-/// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
-/// # let node = p2panda::spawn().await?;
-/// let topic = Topic::random();
-///
-/// let (_tx, mut rx) = node.stream::<String>(topic).await?;
-///
-/// while let Some(stream_event) = rx.next().await {
-///     // .. react to topic stream events
-/// }
-/// #
-/// # Ok(())
-/// # }
-/// ```
-#[derive(Debug)]
-pub struct StreamSubscription<M> {
-    topic: Topic,
-    store: SqliteStore,
-    acked: Acked,
-    stream: ReceiverStream<StreamEvent<M>>,
-}
-
-impl<M> StreamSubscription<M> {
-    /// Associated topic.
-    pub fn topic(&self) -> Topic {
-        self.topic
-    }
-
-    /// Explicitly acknowledge operation.
-    ///
-    /// Fails silently if operation is not known (it might have been pruned, etc.).
-    ///
-    /// If the [`AckPolicy`] is set to "explicit", users want to call this method _after_
-    /// applicaton-level processing has successfully finished. See high-level description in
-    /// [`Node::stream`](crate::node::Node::stream) for more details.
-    pub async fn ack(&self, id: Hash) -> Result<(), AckedError> {
-        if let Some(operation) = OperationStore::<_, _>::get_operation(&self.store, &id).await? {
-            self.acked.ack(&operation.header).await?;
-        }
-
-        Ok(())
-    }
-}
-
-impl<M> Stream for StreamSubscription<M>
-where
-    M: Serialize + for<'a> Deserialize<'a> + Send + 'static,
-{
-    type Item = StreamEvent<M>;
-
-    fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        self.stream.poll_next_unpin(cx)
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.stream.size_hint()
-    }
 }
 
 /// Operations with application messages, system events and errors coming from a topic stream
@@ -1183,29 +877,4 @@ pub enum Source {
 
     /// Source when an operation was published locally or replayed.
     LocalStore,
-}
-
-/// Error occurred when creating operation and publishing it to topic stream.
-#[derive(Debug, Error)]
-pub enum PublishError {
-    #[error("an error occurred while serializing the message for publication: {0}")]
-    MessageEncoding(#[from] EncodeError),
-
-    #[error("an error occurred while creating an operation in the forge: {0}")]
-    Forge(#[from] ForgeError),
-
-    #[error("an error occurred while publishing an operation to the log sync stream: {0}")]
-    SyncHandle(String),
-
-    #[error("could not send to processor pipeline: {0}")]
-    SendToProcessor(String),
-}
-
-#[derive(Debug, Error)]
-pub enum ImportError {
-    #[error("could not send to processor pipeline: {0}")]
-    SendToProcessor(String),
-
-    #[error("an error occurred awaiting message from processor: {0}")]
-    ReceiveFromProcessor(String),
 }
